requirements.txt
diffusers==0.25.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.45.2
numpy==1.26.4
xformers==0.0.22.post7
tokenizers==0.20.1
accelerate
safetensors
einops
spaces==0.19.4
omegaconf
peft
huggingface-hub==0.24.6
opencv-python
insightface
gradio==4.42.0
controlnet_aux
gdown
onnx==1.16.2
onnxruntime-gpu==1.19.0
# SpectralFusion dependencies
PyWavelets>=1.4.1
scipy>=1.11.0
Pillow>=9.5.0

# Evaluation (optional)
lpips>=0.1.4
torchmetrics>=0.11.0
git+https://github.com/openai/CLIP.git
